refactor calcilators page used tanstack query
