+++ src/hooks/calculators/index.js
@@ -4,6 +4,8 @@ export * from "./useAddPensionFund";
 
 // Emergency
 export * from "./useEmergencyFund";
+export * from "./useAddEmergencyFund";
 
 // Invesment
 export * from "./useInvestment";
+export * from "./useAddInvesment";
