+++ src/hooks/calculators/useAddInvesment.js
@@ -1,18 +1,18 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import toast from "react-hot-toast";
 
-import { addPensionFund } from "@/services/calculators-service";
+import { addInvestment } from "@/services/calculators-service";
 
-export const useAddPensionFund = () => {
+export const useAddInvesment = () => {
   const queryClient = useQueryClient();
 
   const { mutate, isPending } = useMutation({
-    mutationFn: async (data) => await addPensionFund(data),
+    mutationFn: async (data) => await addInvestment(data),
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["pensions"] });
+      queryClient.invalidateQueries({ queryKey: ["investment"] });
     },
     onError: (err) => toast.error(err.message),
   });
 
-  return { addPensionFund: mutate, isPending };
+  return { addInvestment: mutate, isPending };
 };
